requirements.txt
streamlit==1.47.1
opencv-python==4.12.0.88
numpy==2.2.6
pillow==11.3.0
